Refactor presto spark launcher
Extract Presto Spark bootstraping code into the PrestoSparkRunner.
Having a runner that can be initialized without providing the ClientOptions
is more flexible, as it is not strongly coupled with the CLI arguments.
